msk-lambda-schema-avro-python-sam/kafka_event_producer_function/requirements.txt
kafka-python>=2.2.14
boto3>=1.34.0
aws-lambda-powertools>=2.0.0
aws-xray-sdk>=2.12.0
